+++ test/pipe.spec.ts
@@ -3,7 +3,7 @@ import { pipe } from '../src/pipe';
 describe('pipe', () => {
   test('should compose two functions and return function.', () => {
     const composedFn = pipe(
-      (n) => `${n} is number`,
+      (n: number) => `${n} is number`,
       (n) => n + '!'
     );
 
@@ -15,7 +15,7 @@ describe('pipe', () => {
       'array of provided arguments of composed function.',
     () => {
       // @ts-expect-error
-      const composedFn = pipe();
+      const composedFn: any = pipe();
 
       expect(composedFn(6, 5, 4)).toEqual([6, 5, 4]);
     }
@@ -34,5 +34,14 @@ describe('pipe', () => {
     const composed = pipe(async (s: string) => s, parseInt);
 
     expect(await composed('1')).toBe(1);
+  });
+
+  it('should compose functions with one mandatory parameter and with variadic parameter', () => {
+    const composed = pipe(
+      (n: number) => n,
+      (...args: Array<number>) => args.reduce((a, c) => a + c, 0)
+    );
+
+    expect(composed(1)).toBe(1);
   });
 });