+++ src/once.ts
@@ -1,33 +1,18 @@
-import { isNothing } from './is_just_nothing';
-import { maybe, none, Option } from './option';
-
-interface OnceFunction {
-  <T extends ReadonlyArray<unknown>, R>(fn: (...args: T) => R): (
-    ...args: T
-  ) => Option<R>;
-  <T extends ReadonlyArray<unknown>, R>(
-    fn: (...args: T) => R,
-    after: (...args: T) => R
-  ): (...args: T) => R;
-}
-
 /**
  * Invokes _fn_ only once.
  * If _after_ function is provided, then
  * it will be called after _fn_ has been executed.
  */
-export const once: OnceFunction = <T extends ReadonlyArray<unknown>, R>(
-  fn: (...args: T) => R,
-  after?: (...args: T) => R
-) => {
-  let done = false;
+export const once = <F extends (...args: ReadonlyArray<unknown>) => unknown>(
+  fn: F,
+  after?: F
+): F => {
+  let result: ReturnType<F>;
+  let _fn = fn;
 
-  return (...args: T) => {
-    if (done) {
-      return isNothing(after) ? none : after(...args);
-    } else {
-      done = true;
-      return isNothing(after) ? maybe(fn(...args)) : fn(...args);
-    }
-  };
+  return ((...args) => {
+    result = _fn(...args) as ReturnType<F>;
+    _fn = after ?? ((() => result) as F);
+    return result;
+  }) as F;
 };