+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Added
 
-- `demethodize`, `binary` function.
+- `demethodize`, `binary`, `flip` function.
 - Placeholder value (`_`) to preserve place for argument.
+- `Reverse` type.
 
 ### Changed